fix slack lib
